[refactor] part / part of 적용
